requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
faker
pytest
pytest-cov
httpx
python-dotenv